+++ src/main/java/wooteco/subway/controller/web/line/LineController.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line.controller;
+package wooteco.subway.controller.web.line;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.line.domain.Line;
-import wooteco.subway.line.exception.WrongLineIdException;
 import wooteco.subway.line.exception.WrongLineInformationException;
 import wooteco.subway.line.service.LineService;
 import wooteco.subway.section.domain.OrderedSections;
@@ -41,7 +40,7 @@ public class LineController {
         }
         Line line = new Line(lineRequest.getName(), lineRequest.getColor());
         Line newLine = lineService.save(line);
-        sectionService.create(newLine.getId(), lineRequest.getUpStationId(), lineRequest.getDownStationId(), lineRequest.getDistance());
+        sectionService.add(newLine.getId(), lineRequest.getUpStationId(), lineRequest.getDownStationId(), lineRequest.getDistance());
 
         LineResponse lineResponse = new LineResponse(newLine.getId(), newLine.getName().text(), newLine.getColor().text());
         return ResponseEntity.created(URI.create("/lines/" + newLine.getId())).body(lineResponse);
@@ -74,7 +73,7 @@ public class LineController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
         Line line = lineService.findById(id);
-        sectionService.deleteLine(id);
+        sectionService.removeLine(id);
         lineService.delete(line);
         return ResponseEntity.noContent().build();
     }
