+++ src/test/java/wooteco/subway/line/controller/LineControllerTestUtils.java
@@ -4,6 +4,7 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.springframework.http.MediaType;
+import wooteco.subway.controller.web.line.LineRequest;
 
 public class LineControllerTestUtils {
     public static ExtractableResponse<Response> 지하철노선을_생성한다(LineRequest requestBody) {
