+++ src/test/java/wooteco/subway/station/controller/StationControllerTestUtils.java
@@ -4,6 +4,7 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.springframework.http.MediaType;
+import wooteco.subway.controller.web.station.StationRequest;
 
 public class StationControllerTestUtils {
     public static ExtractableResponse<Response> 지하철역을_생성한다(StationRequest stationRequest) {
